Prevent deletion of nonexistent components
To resolve the issue of the delete button being present on components
that do not exist or on the new component page, this commit hides the
delete button on such pages. If the user tries to edit a component that
does not exist, the app takes them to the permission denied.

Bug: T240237
Change-Id: I55cb1809dbc5e0e979e8ebced14880638f4c0f0d