Fix margin and padding for alerts, headings, labels and buttons
This commit fixed the margin and paddings for alerts, headings, labels
and buttons. It fixes the alerts margin by adding a margin above it
separating it from the top navbar. It fixes the headings/label margin
by making the margin equal above and below them so inputs and other
elements don't get too close the labels and headings. It also
makes the "add" and "confirm" buttons take up the entire width of the
screen like the other buttons. It also adds a wrapper to the submit
button in preferences.html and changes the button size for smaller
devices in edit.html since it doesn't have much text.

Change-Id: I44c32a2c841965f90b7de026b41d207ddf3789fe